feat(ui): EstimatePanel with version subnav; job-scoped estimate routes + shim
Extracts EstimateDetailPage's guts into EstimatePanel (job-scoped, no
job/contact fetch of its own), adds a DocSubnav strip built from the job's
estimate versions plus its change orders, and introduces
/jobs/:jobId/estimate[/:docId] with URL-param -> remembered -> latest
document resolution (JobEstimatePage). The old /estimates/:id route becomes
a thin redirect shim that resolves the job and forwards to the new
job-scoped URL.
